--- src/domains/transaction/components/RecipientsList/index.ts
@@ -1,1 +1,0 @@
-export * from "./RecipientsList";
